third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/origin-exact-matching/04-expected.txt



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Check that exact matching is used when comparing a request's originating url and the value provided by Access-Control-Allow-Origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Should allow origin: ' http://localhost:8000'
PASS xhr.send(null) is undefined.
PASS successfullyParsed is true

TEST COMPLETE
